server/cafe-backend/pyproject.toml
[project]
name = "cafe-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-bridge>=0.3",
    "django>=5.1.4",
    "nanoid>=2.0.0",
    "django-allauth[socialaccount]>=65.3.1",
    "python-dotenv>=1.0.1",
    "django-hijack>=3.7.1",
    "debugpy>=1.8.11",
    "rules>=3.5",
    "names>=0.3.0",
    "cryptography>=44.0.0",
    "huey>=2.5.2",
    "redis>=5.2.1",
    "httpx>=0.28.1",
    "meilisearch>=0.34.1",
    "django-simple-history>=3.8.0",
    "typesense>=1.1.1",
    "tenacity>=9.1.2",
    "gunicorn>=23.0.0",
    "sentry-sdk[django]>=2.34.1",
    "pytest-django>=4.11.1",
    "pytest-env>=1.1.5",
    "freezegun>=1.5.5",
    "django-cors-headers>=4.7.0",
    "jsonata-python>=0.6.1",
]
